requirements.txt
requests
beautifulsoup4
whois
python-magic
tweepy
pandas
